requirements.txt
python-binance>=1.0.19
requests>=2.28.0
websocket-client>=1.6.0
